Fixed Windows 64-bit build error.
git-svn-id: http://wxmax.googlecode.com/svn/trunk/wx.mod@862 971f8453-5c39-0410-a452-3998bf019487

+++ wxrarinputstream.mod/src/unrar/crc.cpp
@@ -20,7 +20,7 @@ uint CRC(uint StartCRC,const void *Addr,uint Size)
     InitCRC();
   byte *Data=(byte *)Addr;
 #if defined(LITTLE_ENDIAN) && defined(PRESENT_INT32) && defined(ALLOW_NOT_ALIGNED_INT)
-  while (Size>0 && ((long)Data & 7))
+  while (Size>0 && ((uintptr_t)Data & 7))
   {
     StartCRC=CRCTab[(byte)(StartCRC^Data[0])]^(StartCRC>>8);
     Size--;
